backend/embedding_service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
sentence-transformers
psycopg2-binary
pgvector
fastapi
uvicorn
